refactored my interface
