+++ src/main/java/com/saltsoftware/repository/employee/EmployeeRepository.java
@@ -1,4 +1,4 @@
-package com.saltsoftware.repository.employee.impl;
+package com.saltsoftware.repository.employee;
 
 import com.saltsoftware.entity.employee.Employee;
 import com.saltsoftware.repository.Repository;
