+++ src/main/java/com/saltsoftware/repository/employee/impl/EmployeeRepositoryImpl.java
@@ -1,5 +1,7 @@
 package com.saltsoftware.repository.employee.impl;
 import com.saltsoftware.entity.employee.Employee;
+import com.saltsoftware.repository.employee.EmployeeRepository;
+
 import java.util.HashSet;
 import java.util.Set;
 /**
